requirements.txt
streamlit
openai
langchain
langchain-openai
python-dotenv
requests
fastapi
uvicorn
